Move congestion code to congestion.c
