+++ lib/fragment.c
@@ -561,27 +561,6 @@ int quicrq_fragment_is_ready_to_send(void* v_media_ctx, size_t data_max_size, ui
     return is_ready;
 }
 
-int quicrq_fragment_evaluate_backlog(quicrq_fragment_publisher_context_t* media_ctx)
-{
-    int has_backlog = 0;
-    const uint64_t backlog_threshold = 5;
-
-    if (media_ctx->current_offset > 0 || media_ctx->length_sent > 0) {
-        has_backlog = media_ctx->has_backlog;
-    }
-    else if (media_ctx->current_group_id < media_ctx->cache_ctx->next_group_id ||
-        (media_ctx->current_group_id == media_ctx->cache_ctx->next_group_id &&
-            media_ctx->current_object_id + backlog_threshold < media_ctx->cache_ctx->next_object_id)) {
-        has_backlog = 1;
-        media_ctx->has_backlog = 1;
-    }
-    else {
-        has_backlog = 0;
-        media_ctx->has_backlog = 0;
-    }
-    return has_backlog;
-}
-
 int quicrq_fragment_publisher_fn(
     quicrq_media_source_action_enum action,
     void* v_media_ctx,
@@ -731,23 +710,33 @@ int quicrq_fragment_publisher_fn(
 /* Evaluate whether the media context has backlog, and check
 * whether the current object should be skipped.
 */
+
 int quicrq_fragment_datagram_publisher_object_eval(
     quicrq_stream_ctx_t* stream_ctx,
     quicrq_fragment_publisher_context_t* media_ctx, int* should_skip, uint64_t current_time)
 {
     int ret = 0;
-    const int64_t delta_t_max = 5 * 33333;
+    int has_backlog = 0;
 
     *should_skip = 0;
     if (media_ctx->current_fragment->object_id != 0 &&
         media_ctx->current_fragment->data_length > 0) {
-        if (stream_ctx->cnx_ctx->qr_ctx->quic != NULL &&
-            media_ctx->current_fragment != NULL) {
-            int64_t delta_t = current_time - media_ctx->current_fragment->cache_time;
-            int has_backlog = delta_t > delta_t_max;
+        switch (stream_ctx->cnx_ctx->qr_ctx->congestion_control_mode) {
+        case quicrq_congestion_control_none:
+            break;
+        case quicrq_congestion_control_group:
+            /* TODO: warp mode evaluation */
+            break;
+        case quicrq_congestion_control_delay:
+        default:
+            if (stream_ctx->cnx_ctx->qr_ctx->quic != NULL &&
+                media_ctx->current_fragment != NULL) {
+                has_backlog = quicrq_evaluate_datagram_backlog(media_ctx, current_time);
 
-            *should_skip = quicrq_congestion_check_per_cnx(stream_ctx->cnx_ctx,
-                media_ctx->current_fragment->flags, has_backlog, current_time);
+                *should_skip = quicrq_congestion_check_per_cnx(stream_ctx->cnx_ctx,
+                    media_ctx->current_fragment->flags, has_backlog, current_time);
+            }
+            break;
         }
     }
 
@@ -1177,6 +1166,7 @@ void* quicrq_fragment_publisher_subscribe(quicrq_fragment_cache_t* cache_ctx, qu
     if (media_ctx != NULL) {
         memset(media_ctx, 0, sizeof(quicrq_fragment_publisher_context_t));
         media_ctx->cache_ctx = cache_ctx;
+        media_ctx->congestion_control_mode = stream_ctx->cnx_ctx->qr_ctx->congestion_control_mode;
         if (stream_ctx != NULL) {
             stream_ctx->start_group_id = cache_ctx->first_group_id;
             stream_ctx->start_object_id = cache_ctx->first_object_id;
